Moving private classes from the netcoreapp tests to the regular CollectionTests as they were not available in the netstandard CI test run
